+++ apps/batches/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
-from .views import BatchViewSet, BatchImportView
-
+from .views import BatchViewSet
+from .import_batch import BatchImportView
 
 batch_list = BatchViewSet.as_view({
     "get": "list",
@@ -17,6 +17,6 @@ batch_detail = BatchViewSet.as_view({
 urlpatterns = [
     path("batch/list", batch_list, name="batch-list"),
     path("batch/item/<int:pk>", batch_detail, name="batch-item"),
-    path("batch/export", BatchImportView.as_view(), name="export-batch"),
+    path("batch/import", BatchImportView.as_view(), name="import-batch"),
     
 ]
